--- maestro-orchestra/src/test/resources/workspaces/016_path-aliases/config-in-folder/flows/flow.yaml
@@ -1,4 +1,0 @@
-appId: com.example
----
-- runFlow: '~sf/subflow.yaml'
-- runScript: '~js/script.js'
