--- maestro-orchestra/src/test/resources/workspaces/016_path-aliases/config-in-folder/scripts/script.js
@@ -1,1 +1,0 @@
-console.log('Script executed successfully')
