--- maestro-orchestra/src/test/resources/workspaces/016_path-aliases/config-in-folder/subflows/subflow.yaml
@@ -1,3 +1,0 @@
-appId: com.example
----
-- assertTrue: ${true}
