--- maestro-orchestra/src/test/resources/workspaces/016_path-aliases/config-in-root/subflows/subflow.yaml
@@ -1,3 +1,0 @@
-appId: com.example
----
-- assertTrue: ${true}
