+++ maestro-orchestra/src/test/resources/workspaces/016_path-aliases_config-in-folder/workspace/.maestro/config.yaml
