+++ maestro-orchestra/src/test/resources/workspaces/016_path-aliases_config-in-folder/workspace/scripts/script.js
@@ -1,0 +1,2 @@
+console.log('Script executed successfully')
+output.scriptExecuted = true
