+++ maestro-orchestra/src/test/resources/workspaces/016_path-aliases_config-in-root/workspace/config.yaml
