+++ maestro-orchestra/src/test/resources/workspaces/016_path-aliases_config-in-root/workspace/flows/flow.yaml
@@ -1,0 +1,5 @@
+appId: com.example
+---
+- evalScript: ${output.subflowExecuted = false}
+- runFlow: '~sf/subflow.yaml'
+- assertTrue: ${output.subflowExecuted}
