+++ maestro-orchestra/src/test/resources/workspaces/016_path-aliases_config-in-root/workspace/subflows/subflow.yaml
@@ -1,0 +1,4 @@
+appId: com.example
+---
+- assertTrue: ${true}
+- evalScript: ${output.subflowExecuted = true}
