+++ maestro-orchestra/src/test/resources/workspaces/017_js-interpolation/workspace/flows/test-runflow-js-interpolation.yaml
@@ -1,7 +1,7 @@
 appId: com.example
 ---
 # Test runFlow with JS interpolation in file path
-- evalScript: output.flowName = "login"
+- evalScript: ${output.flowName = "login"}
 - runFlow: ~sf/${output.flowName}.yaml
 - assertTrue: ${output.loginFlowExecuted}
