--- maestro-orchestra/src/test/resources/workspaces/017_js-interpolation/workspace/scripts/helper.js
@@ -1,3 +1,0 @@
-output.helperExecuted = true;
-console.log('Helper script executed');
-
