--- maestro-orchestra/src/test/resources/workspaces/017_js-interpolation/workspace/subflows/auth.yaml
@@ -1,5 +1,0 @@
-appId: com.example
----
-- assertTrue: ${true}
-- evalScript: output.authFlowExecuted = true
-
