+++ maestro-orchestra/src/test/resources/workspaces/017_js-interpolation/workspace/subflows/login.yaml
@@ -1,5 +1,5 @@
 appId: com.example
 ---
 - assertTrue: ${true}
-- evalScript: output.loginFlowExecuted = true
+- evalScript: ${output.loginFlowExecuted = true}
